requirements.txt
Flask==0.10.1
future==0.16.0
requests
beautifulsoup4
flask-cors
matplotlib
numpy
opencv-contrib-python
Pillow
tqdm